Alerts con sweet alert librery

+++ AprendiendoAngular/src/app/components/article/article.component.ts
@@ -3,6 +3,7 @@ import { ArticleService } from 'src/app/services/article.service';
 import { Article } from 'src/app/models/article';
 import { Router, ActivatedRoute, Params } from '@angular/router';
 import { Global } from 'src/app/services/global';
+import swal from 'sweetalert';
 
 @Component({
   selector: 'app-article',
@@ -43,16 +44,36 @@ export class ArticleComponent implements OnInit {
   }
 
   delete(id: any) {
-    //utilizo el servicio de articulos al metodo delete, y me subscribo al observable
-    this._articleService.delete(id).subscribe({
-      next: (response) => {
-        //utilizar el servicio y luego lo redirijo al blog
-        this._router.navigate(['/blog']);
-      },
-      error: (error) => {
-        console.log(error);
-        this._router.navigate(['/blog']);
-      },
+    swal({
+      title: '¿Estas seguro?',
+      text: 'Una vez borrado, no podras recuperar el archivo!',
+      icon: 'warning',
+      buttons: [true, true],
+      dangerMode: true
+    })//Se ejecuta la promesa
+    .then((willDelete) => {
+      if (willDelete) {
+        //utilizo el servicio de articulos al metodo delete, y me subscribo al observable
+        this._articleService.delete(id).subscribe({
+          next: (response) => {
+            //Alerta de borrado con exito
+            swal('El artículo ha sido borrado!', {
+              icon: 'success',
+            });
+            //utilizar el servicio y luego lo redirijo al blog
+            this._router.navigate(['/blog']);
+          },
+          error: (error) => {
+            console.log(error);
+            //Alerta de salvar archivo
+        swal('Error, todo sigue igual!');
+
+            this._router.navigate(['/blog']);
+          },
+        });
+      } else {
+        swal('Cancelado, todo sigue igual!');
+      }
     });
   }
 }
